requirements.txt
requests==2.31.0
nltk==3.8.1
spacy==3.7.2
sumy==0.11.0
streamlit==1.28.1
pandas==2.1.3
python-dotenv==1.0.0
scikit-learn==1.3.2
numpy==1.24.3
wordcloud==1.9.2